+++ xbmc/cores/DllLoader/dll_util.cpp
@@ -57,16 +57,16 @@ uintptr_t create_dummy_function(const char* strDllName, const char* strFunctionN
   memcpy(pData, dummy_func, iFunctionSize);
 
   // insert pointers to datapointers into assembly code (fills 0x00000000 in dummy_func)
-  *(int*)(pData + 4) = (int)offDataPointer1;
-  *(int*)(pData + 10) = (int)offDataPointer2;
-  *(int*)(pData + 17) = (int)offDataPointer3;
+  *(int*)(pData + 4) = (intptr_t)offDataPointer1;
+  *(int*)(pData + 10) = (intptr_t)offDataPointer2;
+  *(int*)(pData + 17) = (intptr_t)offDataPointer3;
 
   // 2 fill datapointer with pointer to 5 (string)
-  *(int*)offDataPointer1 = (int)offStringFunc;
+  *(int*)offDataPointer1 = (intptr_t)offStringFunc;
   // 3 fill datapointer with pointer to 6 (string)
-  *(int*)offDataPointer2 = (int)offStringDll;
+  *(int*)offDataPointer2 = (intptr_t)offStringDll;
   // 4 fill datapointer with pointer to dll_dummy_output
-  *(int*)offDataPointer3 = (int)dll_dummy_output;
+  *(int*)offDataPointer3 = (intptr_t)dll_dummy_output;
 
   // copy arguments to 5 (string) and 6 (string)
   memcpy(offStringDll, strDllName, iDllNameSize);
